requirements.txt
numpy
pandas
opencv-python
scikit-learn
matplotlib
seaborn
tensorflow>=2.10
joblib
streamlit
pillow
tqdm
